feat(fe-next-shadcn): group dashboard component
