+++ presets/fe-next-shadcn/app/admin/layout.tsx
@@ -1,12 +1,14 @@
-import { Dashboard } from "shared/components/AdminDashboard";
 import AdminWrapper from "shared/components/AdminWrapper";
+import { Layout } from "shared/components/dashboard";
 
 export default function AdminLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  return <AdminWrapper>
-    <Dashboard></Dashboard>
-  </AdminWrapper>;
+  return (
+    <AdminWrapper>
+      <Layout>{children}</Layout>
+    </AdminWrapper>
+  );
 }
