+++ presets/fe-next-shadcn/shared/components/dashboard/index.ts
@@ -1,0 +1,4 @@
+export { default as Header } from "./Header";
+export { default as ContentWrapper } from "./ContentWrapper";
+export { default as Layout } from "./Layout";
+export { default as Sidebar } from "./Sidebar";
