+++ presets/fe-next-shadcn/shared/hooks/error.hook.ts
@@ -5,7 +5,6 @@ import { errorStore } from "shared/stores"
 export const useHandleError = ()=>{
   const error = useAtomValue(errorStore);
   useEffect(()=>{
-    console.log({error})
     if(error.error){
       alert(error.error.message)
     }
