+++ presets/fe-next-shadcn/shared/utils/api.utils.ts
@@ -1,11 +1,9 @@
 import config from "config";
-import { Api } from "../api/Api";
+import { Api } from "shared/api/Api";
 import Cookie from "js-cookie";
-import { QueryClient } from "react-query";
 import { AxiosError } from "axios";
 import { store, errorStore } from "shared/stores";
 
-export const queryClient = new QueryClient();
 export const api = () => {
     const storage = Cookie.get(config.authStore) || "{}";
     const token = JSON.parse(storage)?.token;
